Implement "soft" deconfig
A few configuration settings cannot be deconfig'ed without breaking
stuff. And some times one needs a default. Implement @ syntax that
allows for the entry to exist in the export, but only be used in the
case of no entry in the active storage.

The implementation isn't super pretty at the moment.